+++ backend/build.gradle
@@ -33,7 +33,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-elasticsearch'
     implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.springframework.security:spring-security-jwt:1.0.9.RELEASE'
-    implementation 'org.springframework.security.oauth:spring-security-oauth2:2.3.3.RELEASE'
+    implementation 'org.springframework.security.oauth:spring-security-oauth2:2.4.0.RELEASE'
 
     testImplementation('org.springframework.boot:spring-boot-starter-test')
 //    testImplementation('org.springframework.security:spring-security-test')
@@ -54,7 +54,6 @@ jacocoTestReport {
         html.enabled = true
         csv.enabled = false
         html.destination file("coverage/html")
-
     }
 }
