--- backend/src/main/java/com/raven/croaker/service/internal/UserServiceImpl.java
@@ -1,37 +1,0 @@
-package com.raven.croaker.service.internal;
-
-import com.raven.croaker.domain.User;
-import com.raven.croaker.domain.UserRepository;
-import com.raven.croaker.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserServiceImpl implements UserService {
-    private UserRepository userRepository;
-
-    @Autowired
-    public void setUserRepository(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public User save(User user) {
-        return userRepository.save(user);
-    }
-
-    @Override
-    public User findByUsername(String username) {
-        return userRepository.findByUsername(username);
-    }
-
-    @Override
-    public User findByEmail(String email) {
-        return userRepository.findByEmail(email);
-    }
-
-    @Override
-    public Iterable<User> findAll() {
-        return userRepository.findAll();
-    }
-}
